requirements-dev.txt
ansible-core>=2.17,<2.19
ansible-lint>=24.0.0
pytest>=8.0.0
yamllint>=1.35.0
antsibull-changelog>=0.32.0
